+++ crdesigner/map_conversion/opendrive/opendrive_parser/elements/roadPlanView.py
@@ -258,7 +258,7 @@ class PlanView:
             geo_idx = np.arange(self._geo_lengths.shape[0])[mask][sub_idx] - 1
         except ValueError:
             # s_pos is after last geometry because of rounding error
-            if np.isclose(s_pos, self._geo_lengths[-1], 0.001, 0.001):
+            if np.isclose(s_pos, self._geo_lengths[-1], 0.01, 0.01):  # todo parameter
                 geo_idx = self._geo_lengths.size - 2
             else:
                 raise Exception(